backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
aiofiles>=24.1.0
sqlalchemy>=2.0.36
pydantic>=2.9.0
pydantic-settings>=2.6.0
opencv-python-headless>=4.10.0
easyocr>=1.7.2
paddlepaddle>=2.6.0
paddleocr>=2.7.0
pillow>=11.0.0
numpy>=2.1.0
spacy>=3.8.0
transformers>=4.47.0
torch>=2.5.0
torchvision>=0.20.0
tokenizers>=0.21.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
httpx>=0.28.0
langdetect>=1.0.9
python-docx>=1.1.2
bcrypt>=4.2.0
pyjwt>=2.10.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
# pdf2image for PDF processing
pdf2image>=1.17.0
PyPDF2>=3.0.0
# For enhanced image preprocessing
scikit-image>=0.24.0
# For watermarking
scipy>=1.14.0
# For encrypted SQLite (optional)
# pysqlcipher3>=1.2.0
# Optional on Apple Silicon for local handwritten VLM transcription:
# mlx-vlm>=0.1.0
